+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to
 
 ### Added
 
-- Added `TenableClient.iteraetAssets()` method. Made other asset export
+- Added `TenableClient.iterateAssets()` method. Made other asset export
   endpoints private.
 
 ## 7.1.4 - 2021-07-13